requirements_all.txt
aiofiles
aiohttp
aiosignal
altair
anyio
argilla
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-timeout
attrs
backcall
backoff
beautifulsoup4
bleach
bokeh
Brotli
certifi
cffi
chardet
charset-normalizer
chromadb
click
clickhouse-connect
codeboxapi
codeinterpreterapi
coloredlogs
comm
commonmark
contourpy
cryptography
cycler
dataclasses-json
debugpy
decorator
defusedxml
Deprecated
docarray
duckdb
entrypoints
et-xmlfile
exceptiongroup
executing
fastapi
fastjsonschema
ffmpy
filelock
filetype
flatbuffers
fonttools
fqdn
frozenlist
fsspec
gradio
gradio_client
greenlet
h11
hnswlib
httpcore
httptools
httpx
huggingface-hub
humanfriendly
idna
install
ipykernel
ipython
ipython-genutils
isoduration
jedi
Jinja2
joblib
jsonpointer
jsonschema
jupyter-events
jupyter-kernel-gateway
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab-pygments
kiwisolver
langchain
langchainplus-sdk
langsmith
lark
linkify-it-py
lxml
lz4
Markdown
markdown-it-py
MarkupSafe
marshmallow
marshmallow-enum
matplotlib
matplotlib-inline
mdit-py-plugins
mdurl
mistune
monotonic
mpmath
msg-parser
multidict
mutagen
mypy-extensions
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
nltk
notebook
notebook_shim
numexpr
numpy
olefile
onnxruntime
openai
openapi-schema-pydantic
openpyxl
opensearch-py
orjson
overrides
packaging
pandas
pandocfilters
panel
param
parso
pdf2image
pdfminer.six
pexpect
pickleshare
Pillow
platformdirs
posthog
prometheus-client
prompt-toolkit
protobuf
psutil
ptyprocess
pulsar-client
pure-eval
pycparser
pycryptodomex
pydantic
pydub
Pygments
pypandoc
pyparsing
pypdf
pyrsistent
python-dateutil
python-docx
python-dotenv
python-json-logger
python-magic
python-multipart
python-pptx
pytz
pyviz-comms
PyYAML
pyzmq
regex
requests
rfc3339-validator
rfc3986
rfc3986-validator
rich
scikit-learn
scipy
semantic-version
Send2Trash
six
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
sympy
tabulate
tenacity
terminado
threadpoolctl
tiktoken
tinycss2
tokenizers
toolz
tornado
tqdm
traitlets
typer
types-requests
types-urllib3
typing-inspect
typing_extensions
tzdata
uc-micro-py
unstructured
uri-template
urllib3
uvicorn
uvloop
watchfiles
wcwidth
webcolors
webencodings
websocket-client
websockets
wikipedia
wrapt
xlrd
XlsxWriter
xyzservices
yarl
yt-dlp
zstandard
